requirements.txt
Flask
SQLAlchemy
python-telegram-bot
requests
flask_cors